[refactor] Cleanup imports in tests to be src, export AnyPublicKey/Signature
