+++ src/core/crypto/index.ts
@@ -7,3 +7,5 @@ export * from "./multiEd25519";
 export * from "./secp256k1";
 export * from "./multiKey";
 export * from "./hdKey";
+export * from "./anyPublicKey";
+export * from "./anySignature";
