+++ tests/e2e/transaction/transactionArguments.test.ts
@@ -21,6 +21,9 @@ import {
   isMultiAgentSignature,
   isFeePayerSignature,
   isUserTransactionResponse,
+  MoveOption,
+  MoveString,
+  MoveVector,
 } from "../../../src";
 import {
   MAX_U128_BIG_INT,
@@ -30,7 +33,6 @@ import {
   MAX_U64_BIG_INT,
   MAX_U8_NUMBER,
 } from "../../../src/bcs/consts";
-import { MoveOption, MoveString, MoveVector } from "../../../src/bcs/serializable/moveStructs";
 import {
   fundAccounts,
   rawTransactionHelper,
